Fixed Grid Changed Viscosity Algorithm Changed Double Density Relaxation Algorithm Added Zoom

+++ include/FluidSim/Particle.h
@@ -9,7 +9,7 @@ public:
     glm::vec2 pos;
     glm::vec2 ppos;
     glm::vec2 vel;
-    unsigned int index;
+    float index;
 };
 
 #endif //FLUIDSIM_PARTICLE_H
